Add unit tests for add survey
