+++ ras_party/controllers/queries.py
@@ -61,5 +61,5 @@ def query_business_respondent_by_respondent_id_and_business_id(business_id, resp
     """
     logger.debug('Querying business respondent with respondent id {}'.format(respondent_id))
 
-    respondent_id = str(respondent_id)
-    return session.query(BusinessRespondent).filter(and_(str(BusinessRespondent.respondent_id) == respondent_id, BusinessRespondent.business_id == business_id)).first()
+    response = session.query(BusinessRespondent).filter(and_(BusinessRespondent.business_id == business_id, BusinessRespondent.respondent_id == respondent_id)).first()
+    return response
